LG: replace ushort/ulong/... with uint16_t/uint32_t/...

+++ ShockMac/Libraries/LG/Source/lgsprntf.c
@@ -6,15 +6,15 @@ This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation, either version 3 of the License, or
 (at your option) any later version.
- 
+
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
- 
+
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
- 
+
 */
 // lgsprntf.c -- no-floats baby sprintf
 
@@ -26,38 +26,38 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 * $Log: lgsprntf.c $
  * Revision 1.12  1994/08/15  16:52:29  tjs
  * length specifiers 'l' & 'h' supported, as far as that goes.
- * 
+ *
  * Revision 1.11  1994/08/15  13:12:33  jak
  * Don't write though null pointer.
- * 
+ *
  * Revision 1.10  1994/08/14  15:48:51  tjs
  * Precision support for strings.
- * 
+ *
  * Revision 1.9  1994/08/12  02:27:56  tjs
  * Let's try that again.
- * 
+ *
  * Revision 1.8  1994/08/12  02:17:32  tjs
  * %x=%p
- * 
+ *
  * Revision 1.7  1994/02/27  02:25:58  tjs
  * Changed format to const
- * 
+ *
  * Revision 1.6  1994/02/01  05:57:36  tjs
  * Added binary numbers, optimization for small decimal integers.
- * 
+ *
  * Revision 1.5  1993/12/17  12:10:16  tjs
  * Fixed bug with %%
- * 
+ *
  * Revision 1.4  1993/11/04  14:08:17  tjs
  * Added alternate form for bools.
- * 
+ *
  * Revision 1.3  1993/11/04  11:46:11  tjs
  * Added error message for %S if function not installed, precision
  * handling for fixed-point numbers.
- * 
+ *
  * Revision 1.2  1993/11/04  09:42:35  tjs
  * Fixed bug with decimal point placement in fixed-point numbers.
- * 
+ *
  *
  *
  *
@@ -73,12 +73,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define MAX_FIX_PRECIS 4
 
 static int int_to_str(char *buf, int val);
-static int uint_to_str(char *buf, uint val, int base, char alph);
+static int uint_to_str(char *buf, uint32_t val, int base, char alph);
 
 static char *boolstring[] = {"FALSE","TRUE"};
 static int pten[MAX_FIX_PRECIS+1]={ 1, 10, 100, 1000, 10000 };
 
-static char *(*sprintf_str_func)(ulong strnum)=NULL;
+static char *(*sprintf_str_func)(uint32_t strnum)=NULL;
 
 // For small positive integers, just indirect into the big
 // magic array and copy the two bytes you find there.  This
@@ -110,10 +110,10 @@ typedef enum { SMALL, DEFAULT, BIG } bigness;
 
 // lg_sprintf()
 // should have the save behaviour as sprintf(), but supports only the
-// %d, %u, %s, %c, %x, %X, %o, %n, and %% conversion characters, and does not 
+// %d, %u, %s, %c, %x, %X, %o, %n, and %% conversion characters, and does not
 // support the space and + flags.
-// In addition, instead of supporting floats, we support formatting 
-// of fixes and fix24's, using the conversion characters %f and %F, 
+// In addition, instead of supporting floats, we support formatting
+// of fixes and fix24's, using the conversion characters %f and %F,
 // respectively.  These are currently always formatted with four digits
 // after the decimal place.  Also supports %b conversion characters for
 // boolean values, formatting them as "TRUE" or "FALSE".  For those of you
@@ -145,7 +145,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
    fix24 arg_fix24;
    bool arg_bool;
    char *arg_str;
-   uint arg_uint;
+   uint32_t arg_uint;
    int arg_int;
    char fix_frac_buf[5];
 
@@ -222,7 +222,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                arg_int=va_arg(arglist,int);
 #ifdef LGSPF_SMALLINT_OPT
                if(arg_int>=0 && arg_int<100) {
-                  *((ushort*)buf)=((ushort*)digiarray)[arg_int];
+                  *((uint16_t*)buf)=((uint16_t*)digiarray)[arg_int];
                   newchars=(arg_int<10)?1:2;
                }
                else
@@ -232,12 +232,12 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                }
                break;
             case 'u': // unsigned int
-               arg_uint=va_arg(arglist,uint);
+               arg_uint=va_arg(arglist,uint32_t);
                if(big==SMALL)
                   arg_uint &= USHRT_MAX;
 #ifdef LGSPF_SMALLINT_OPT
                if(arg_uint<100) {
-                  *((ushort*)buf)=((ushort*)digiarray)[arg_uint];
+                  *((uint16_t*)buf)=((uint16_t*)digiarray)[arg_uint];
                   newchars=(arg_uint<10)?1:2;
                }
                else
@@ -247,7 +247,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                }
                break;
             case 'B': // binary
-               arg_uint=va_arg(arglist,uint);
+               arg_uint=va_arg(arglist,uint32_t);
                buf[dest_ind]='0';
                buf[dest_ind+1]='b';
                newchars=2+uint_to_str(buf+dest_ind+2,arg_uint,2,0);
@@ -256,7 +256,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                src_char='X';
             case 'x':
             case 'X':
-               arg_uint=va_arg(arglist,uint);
+               arg_uint=va_arg(arglist,uint32_t);
                if(big==SMALL)
                   arg_uint &= USHRT_MAX;
                if(altform && arg_uint!=0) {
@@ -268,7 +268,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                   newchars=uint_to_str(buf+dest_ind,arg_uint,16,src_char-'X'+'A');
                break;
             case 'o': // unsigned, octal
-               arg_uint=va_arg(arglist,uint);
+               arg_uint=va_arg(arglist,uint32_t);
                if(big==SMALL)
                   arg_uint &= USHRT_MAX;
                if(altform) {
@@ -276,7 +276,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                   newchars=1+uint_to_str(buf+dest_ind+1,arg_uint,8,0);
                }
                else
-                  newchars=uint_to_str(buf+dest_ind,va_arg(arglist,uint),8,0);
+                  newchars=uint_to_str(buf+dest_ind,va_arg(arglist,uint32_t),8,0);
                break;
             case 'c': // char
                buf[dest_ind]=(unsigned char)va_arg(arglist,int);
@@ -297,7 +297,7 @@ int lg_vsprintf(char *buf, const char *format, va_list arglist)
                break;
             case 'S': // string number
                if(sprintf_str_func) {
-                  arg_str=sprintf_str_func(va_arg(arglist,ulong));
+                  arg_str=sprintf_str_func(va_arg(arglist,uint32_t));
                   goto string_copy;
                }
                else {
@@ -353,7 +353,7 @@ string_copy:
                   buf[dest_ind+newchars]='.';
                   newchars++;
                }
-               break;               
+               break;
             }
 
          if(isalpha(src_char) && !this_is_len) {
@@ -414,7 +414,7 @@ string_copy:
 // a char* (like the resource system's RefGet(), by some coincidence)
 // and you are responsible for any necessary memory management for the
 // string it returns.
-void lg_sprintf_install_stringfunc(char *(*func)(ulong strnum))
+void lg_sprintf_install_stringfunc(char *(*func)(uint32_t strnum))
 {
    sprintf_str_func=func;
 }
@@ -424,7 +424,7 @@ void lg_sprintf_install_stringfunc(char *(*func)(ulong strnum))
 
 // ==== static functions follow =======================
 
-// private local functions used for writing integers and uints into 
+// private local functions used for writing integers and uints into
 // strings.  The int version calls the uint version after doing any
 // necessary setup for negative numbers.  The uint version does the
 // real work, writing the integer into the string in reverse digit
@@ -456,14 +456,14 @@ static int int_to_str(char *buf, int val)
          }
          buf[2+len]='\0';
          return(2+len);
-      }         
+      }
       return(1+int_to_str(buf+1,-val));
    }
 
-   return(uint_to_str(buf,(uint)val,10,0));
+   return(uint_to_str(buf,(uint32_t)val,10,0));
 }
 
-static int uint_to_str(char *buf, uint val, int base, char alph)
+static int uint_to_str(char *buf, uint32_t val, int base, char alph)
 {
    int ind, rev;
    char tmp;
